Refactor utils and constants
