+++ src/Game.cpp
@@ -1,4 +1,4 @@
-#include "SFML/System/Vector2.hpp"
+#include <SFML/System/Vector2.hpp>
 
 #include "Game.h"
