+++ src/constants.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <string>
+
+constexpr auto WINDOW_SIZE = 512u;
+constexpr auto THIRD_OF_WINDOW = WINDOW_SIZE / 3.f;
+const std::string TITLE = "Tic Tac Toe";
